\name{NormqPCR-package}
\alias{NormqPCR-package}
\alias{NormqPCR}
\docType{package}
\title{
  Functions for normalisation of RT-qPCR data.
}
\description{
Functions for normalisation of real-time quantitative PCR data.
}
\details{
\tabular{ll}{
Package: \tab NormqPCR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-10-20\cr
Depends: \tab R(>= 2.7.0), stats, RColorBrewer\cr
License: \tab GPL (>=2)\cr
}
require(NormqPCR)
}
\author{ Matthias Kohl, James Perkins

Maintainer: Matthias Kohl  \email{Matthias.Kohl@stamats.de}}
\references{
  Jo Vandesompele, Katleen De Preter, Filip Pattyn et al. (2002). Accurate 
  normalization of real-time quantitative RT-PCR data by geometric averaging 
  of multiple internal control genes. 
  Genome Biology 2002. 3(7):research0034.1-0034.11.
  \url{http://genomebiology.com/2002/3/7/research/0034/}
}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
