\name{normqPCRSet}
\docType{class}
\alias{class:normqPCRSet}
\alias{normqPCRSet}
\alias{normqPCRSet}

\title{
  Class to Contain and Describe normalise qPCR Data
}

\description{
  Container for normalised qPCR Data
}
\section{Creating Objects}{

  \code{new("normqPCRSet")}

}
\section{Slots}{
}
\section{Methods}{
}

\author{James Perkins}

\examples{
}

\keyword{classes}

